requirements.txt
fastapi
uvicorn
pyyaml
pydantic
chromadb
sentence-transformers

